fixed ui issue
