+++ src/app/globals.css
@@ -200,3 +200,43 @@ body {
     0 0,
     13px 23px;
 }
+
+/* Prevent layout breaking for very long words in Tiptap editor and chat messages */
+.tiptap-editor-wrapper,
+.rich-editor {
+  max-width: 100%;
+  min-width: 0;
+  overflow-x: auto;
+  -webkit-overflow-scrolling: touch;
+}
+
+.ProseMirror,
+.tiptap-editor-wrapper .ProseMirror {
+  overflow-wrap: anywhere;
+  word-break: break-word;
+  white-space: pre-wrap;
+  display: block;
+  width: 100%;
+  box-sizing: border-box;
+  min-width: 0;
+}
+
+.tiptap-editor-wrapper .ProseMirror p,
+.tiptap-editor-wrapper .ProseMirror div,
+.tiptap-editor-wrapper .ProseMirror span {
+  overflow-wrap: anywhere;
+  word-break: break-word;
+  white-space: pre-wrap;
+}
+
+/* Prevent the editor content from forcing parent width */
+.tiptap-editor-wrapper .ProseMirror {
+  min-width: 0;
+}
+
+/* Ensure images inserted via editor are responsive */
+.tiptap-image-wrapper img {
+  max-width: 100%;
+  height: auto;
+  display: block;
+}
