+++ src/components/common/TiptapRichEditor.tsx
@@ -329,7 +329,7 @@ export default function RichTextEditor({
     editorProps: {
       attributes: {
         class:
-          "prose prose-sm sm:prose lg:prose-lg xl:prose-2xl mx-auto focus:outline-none p-1"
+          "prose prose-sm sm:prose lg:prose-lg xl:prose-2xl mx-auto focus:outline-none p-1 break-words break-all"
       },
       handleKeyDown: (view, event) => {
         if (
